ints though

+++ src/Wallpaper.cpp
@@ -29,7 +29,7 @@ int main(){
 	//defaults and stuff
 	Wallpaper wallset = {
 		640 * 2,
-		480 * 1.5,
+		(int)(480 * 1.5),
 		60,
 		&buff
 	};
